First draft for Primer_MV, assorted fixes
Ive started working on the machine vision primer I wanted to write.

Added it to the main site under projects and whatnot.

Noticed that the titles were not being displayed. That's because H1 font is white, but so are the blog backgrounds. I moved to usinng H2 and keeping H1 for the darker backgrounds.